- change type from price, long to double - rename SecondProductDto to AddProductDto - add GetCartItemDto - add GetProductDto - return carrito info with two last class

+++ src/main/java/com/KameHouse/ecom/entity/CartItems.java
@@ -33,13 +33,13 @@ public class CartItems {
         this.id = id;
     }
 
-    private Long price;
+    private Double price;
 
-    public Long getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
